fix(crossSim): manage case when SNPs have negative linkage map positions
The simulation algorithm draw the crossing overs position in a uniform distribution from `0` to `chromosome length`. Therefore, crossing over couldn't appear between SNPs with negative linkage map position.

To fix that, instead of using directly the given positions, I first translate the linkage map position of the SNPs (`newPosition = initialPosition - smallest Position`) so that the smallest position now set to 0.

We can do that because only the relative positions of the SNPs is important when sampling the crossing over position.