doc: refact documentation generation function
Some developpement utility functions was in `src/utils.R`, however those
functions are not necessary for the engine, but only for developement
purpose (eg. writeDoc() function that generate the engine's functions
documentation).

I moved those functions out of the sources of the engine in a `tools`
dirctory, and add in the flake some "app" to run them
(eg. `nix run .\#writeDoc`).

Additional points:
- I rebuilt the documentation and result files examples
- Initially, `writeDoc` didn't worked with `nix run` so it has been fixed.
- small bug was also fixed in `createResultExample()`