fix: override `rPackages.vegan` to not depend on pkgs.xvfb-run
On MacOS, this `rPackages.vegan` cannot be build because it is flagged
as a R-package requiring X. This lead to its `buildInputs` to contain
`pkgs.xvfb-run` which cannot be build on MacOS.

This commit override the `rPackages.vegan` to remove this dependency.

Note: I had trouble to manage the overriding of vcf-r dependencies:
I initially used a filter:

vcfR_withoutX = pkgs.rPackages.vcfR.overrideAttrs (oldAttrs: {
  propagatedBuildInputs =
    (builtins.filter (pkg: pkg != pkgs.rPackages.vegan) oldAttrs.propagatedBuildInputs) ++ [vegan_withoutX];
  nativeBuildInputs =
    (builtins.filter (pkg: pkg != pkgs.rPackages.vegan) oldAttrs.nativeBuildInputs) ++ [vegan_withoutX];
});

However, this cannot work on MacOS since the filter will evaluate pkgs.rPackages.vegan
to test the inequality which will break on MacOS.
Instead I explicitly list the dependencies.