refact+fix(readWriteData): use 0 for missing marker effects
Before the reading of marker effects was not consistent and dominance
effects was sometime loaded as NA or as 0.
Now any missing marker effect is loaded as 0.

This also fix a bug when applying blup calculation with marker effects
with several traits where the set of SNP was different between each
traits. For a given trait, the dominance marker effects was loaded as 0 for the markers
exclusive to another trait, but the rest was loaded as NA. This cause
the "progeny blups calculation" check to wrongly raise the error
DOMINANCE_MODEL_FOR_BLUP_ESTIMATION.